Refactor Component.scala
